pkg/endpoint: move identityLabelsChanged to identity.go
Signed-off by: Ian Vernon <ian@cilium.io>
