+++ pkg/endpoint/endpoint.go
@@ -1603,88 +1603,6 @@ func (e *Endpoint) UpdateLabels(owner Owner, identityLabels, infoLabels labels.L
 	}
 }
 
-func (e *Endpoint) identityLabelsChanged(owner Owner, myChangeRev int) error {
-	e.Mutex.RLock()
-	newLabels := e.OpLabels.IdentityLabels()
-	elog := log.WithFields(logrus.Fields{
-		logfields.EndpointID:     e.ID,
-		logfields.IdentityLabels: newLabels,
-	})
-
-	// Since we unlocked the endpoint and re-locked, the label update may already be obsolete
-	if e.identityResolutionIsObsolete(myChangeRev) {
-		e.Mutex.RUnlock()
-		elog.Debug("Endpoint identity has changed, aborting resolution routine in favour of new one")
-		return nil
-	}
-
-	if e.SecurityIdentity != nil &&
-		string(e.SecurityIdentity.Labels.SortedList()) == string(newLabels.SortedList()) {
-
-		e.Mutex.RUnlock()
-		elog.Debug("Endpoint labels unchanged, skipping resolution of identity")
-		return nil
-	}
-
-	// Unlock the endpoint mutex for the possibly long lasting kvstore operation
-	e.Mutex.RUnlock()
-	elog.Debug("Resolving identity for labels")
-
-	identity, _, err := identityPkg.AllocateIdentity(newLabels)
-	if err != nil {
-		err = fmt.Errorf("unable to resolve identity: %s", err)
-		e.LogStatus(Other, Warning, fmt.Sprintf("%s (will retry)", err.Error()))
-		return err
-	}
-
-	e.Mutex.Lock()
-
-	// Since we unlocked the endpoint and re-locked, the label update may already be obsolete
-	if e.identityResolutionIsObsolete(myChangeRev) {
-		e.Mutex.Unlock()
-
-		err := identity.Release()
-		if err != nil {
-			// non fatal error as keys will expire after lease expires but log it
-			elog.WithFields(logrus.Fields{logfields.Identity: identity.ID}).
-				WithError(err).Warn("Unable to release newly allocated identity again")
-		}
-
-		return nil
-	}
-
-	// If endpoint has an old identity, defer release of it to the end of
-	// the function after the endpoint structured has been unlocked again
-	if e.SecurityIdentity != nil {
-		oldIdentity := e.SecurityIdentity
-		defer func() {
-			err := oldIdentity.Release()
-			if err != nil {
-				elog.WithFields(logrus.Fields{logfields.Identity: oldIdentity.ID}).
-					WithError(err).Warn("BUG: Unable to release old endpoint identity")
-			}
-		}()
-	}
-
-	elog.WithFields(logrus.Fields{logfields.Identity: identity.StringID()}).
-		Debug("Assigned new identity to endpoint")
-
-	e.SetIdentity(identity)
-
-	ready := e.SetStateLocked(StateWaitingToRegenerate, "Triggering regeneration due to new identity")
-	if ready {
-		e.ForcePolicyCompute()
-	}
-
-	e.Mutex.Unlock()
-
-	if ready {
-		e.Regenerate(owner, "updated security labels")
-	}
-
-	return nil
-}
-
 // setPolicyRevision sets the policy wantedRev with the given revision.
 func (e *Endpoint) setPolicyRevision(rev uint64) {
 	e.policyRevision = rev
